Reformatted alignment of dependencies in Cargo.lock
This commit makes rearrangements in the Cargo.lock file to improve its readability by maintaining consistent horizontal alignment of the listed dependencies. This is purely a cosmetic change and has no effect on the functionality of the code itself, but it does improve the readability and standardization of the file.