Add sendCommand function and filename to File interface
Implemented a new functionality to send SSH commands by creating a new function named sendCommand. This function takes command and connection as arguments, sends the command, and returns the result as promise. Additionally, the 'filename' property has been added to the 'File' interface. Updated the 'listDirectory' method to filter out files with the names "." and "..".