Refactor Sidebar component and update button behavior
The Sidebar component in the ConnectionsList was refactored to remove the usage of 'props.manager' and switch to using 'ConnectionManager.instance'. The removal and connection actions of list items are now handled directly with ConnectionManager's methods. Also, matching color for tooltip and button for the "Delete" action was added for better user experience in visual feedback. The 'navigate' function from 'react-router-dom' was implemented to replace the window.location.href for better SPA navigation handling.